removed enqueue file and placed code within class-loan-calculator, moved loan-calculator-settings & loan-calculator-admin to admin partials folder, moved loan-calculator-public to public partials folder, created functions within settings file to reduce code length for sematic and DRY purposes
