+++ plugins/loan_calculator/admin/partials/loan-calculator-admin-display.php
@@ -13,4 +13,31 @@
  */
 ?>
 
-<!-- This file should primarily consist of HTML with a little bit of PHP. -->
+<div class="wrap">
+    <h1><?= esc_html(get_admin_page_title()); ?></h1>
+    <?php settings_errors(); ?>
+    <form action="options.php" method="post">
+        <?php 
+        	settings_fields('loan-calculator-settings-group');
+            echo "<div class='settings-group width-50'>";
+            do_settings_sections('interest_rates');
+            echo "</div>";
+        	echo "<div class='settings-group'>";
+        	do_settings_sections('loan_amounts');
+        	echo "</div>";
+            echo "<div class='settings-group'>";
+        	do_settings_sections('loan_terms');
+            echo "</div>";
+            echo "<div class='settings-group'>";
+            do_settings_sections('pop_ups');
+            echo "</div>";
+            echo "<div class='settings-group'>";
+            do_settings_sections('labels');
+            echo "</div>";
+            echo "<div class='settings-group'>";
+            do_settings_sections('links');
+            echo "</div>";
+        	submit_button('Save Settings');
+        ?>
+    </form>
+</div>
